server/requirements-server.txt
absl-py==2.1.0
anaconda-anon-usage @ file:///C:/b/abs_c3w_h1zzjg/croot/anaconda-anon-usage_1710965204622/work
annotated-types==0.7.0
anyio==4.8.0
archspec @ file:///croot/archspec_1709217642129/work
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.3.0
boltons @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/boltons_1699480450092/work
boto3==1.37.0
botocore==1.37.0
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_3d36mno480/croot/brotli-split_1714483178642/work
cachetools==5.5.0
certifi @ file:///C:/b/abs_1fw_exq1si/croot/certifi_1725551736618/work/certifi
cffi @ file:///C:/b/abs_90yq4lmu83/croot/cffi_1726856448345/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
conda @ file:///C:/b/abs_3bs4tei4e9/croot/conda_1733841834375/work
conda-content-trust @ file:///C:/b/abs_bdfatn_wzf/croot/conda-content-trust_1714483201909/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1727775630457/work/src
conda-package-handling @ file:///C:/b/abs_831_pypuuz/croot/conda-package-handling_1718138287790/work
conda_package_streaming @ file:///C:/b/abs_fe6zb0_j8_/croot/conda-package-streaming_1718136098642/work
cryptography @ file:///C:/b/abs_35g500qir4/croot/cryptography_1724940575116/work
databases==0.9.0
distro @ file:///C:/b/abs_71xr36ua5r/croot/distro_1714488282676/work
ecdsa==0.19.0
fastapi==0.115.8
fastjsonschema==2.21.1
ffmpeg==1.4
flatbuffers==24.3.25
frozendict @ file:///C:/b/abs_2alamqss6p/croot/frozendict_1713194885124/work
gast==0.6.0
google-auth==2.37.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
jiter==0.8.2
jmespath==1.0.1
joblib @ file:///C:/b/abs_f4b98l6lgk/croot/joblib_1718217224240/work
jsonpatch @ file:///C:/b/abs_4fdm88t7zi/croot/jsonpatch_1714483974578/work
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupytext==1.16.4
keras==3.7.0
libclang==18.1.1
libmambapy @ file:///C:/b/abs_aam1usey89/croot/mamba-split_1725563515831/work/libmambapy
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
menuinst @ file:///C:/b/abs_76jwd47ck1/croot/menuinst_1723567618942/work
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml-dtypes==0.4.1
namex==0.0.8
nbformat==5.10.4
numexpr @ file:///C:/b/abs_05o8p7bfml/croot/numexpr_1730215959182/work
numpy @ file:///C:/b/abs_c1ywpu18ar/croot/numpy_and_numpy_base_1708638681471/work/dist/numpy-1.26.4-cp312-cp312-win_amd64.whl#sha256=becc06674317799ad0165a939a7613809d0bee9bd328a1e4308c57c39cacf08c
openai==1.61.1
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging @ file:///C:/b/abs_c3vlh0z4jw/croot/packaging_1720101866539/work
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp312-cp312-win_amd64.whl#sha256=2c8058e41dda2e7ecfff0740ad1d98f17d528c740f71b038c523a9d24d2321b0
passlib==1.7.4
platformdirs @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/platformdirs_1701797392447/work
pluggy @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pluggy_1699472504117/work
proto-plus==1.25.0
protobuf==5.29.2
psycopg2==2.9.10
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycosat @ file:///C:/b/abs_5csdern___/croot/pycosat_1714513102923/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
pywin32==308
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests @ file:///C:/b/abs_9frifg92q2/croot/requests_1721410901096/work
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruamel.yaml @ file:///C:/b/abs_0cunwx_ww6/croot/ruamel.yaml_1727980181547/work
ruamel.yaml.clib @ file:///C:/b/abs_5fk8zi6n09/croot/ruamel.yaml.clib_1727769837359/work
s3transfer==0.11.2
scikit-learn @ file:///C:/b/abs_aajc9f7v6r/croot/scikit-learn_1722067652693/work/dist/scikit_learn-1.5.1-cp312-cp312-win_amd64.whl#sha256=96ed6defb86c154d2ff360afdbf0c37c33c4c729d4a15c9d8ada0b70fd817816
scipy @ file:///C:/Users/dev-admin/cr_scipy-release/scipy_1731632527468/work/dist/scipy-1.14.1-cp312-cp312-win_amd64.whl#sha256=34653c6cd617d15590b20aad84b54b5f7f6ed8f683edeeadb2260696234ef19e
setuptools==75.1.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl @ file:///C:/b/abs_def0dwqlft/croot/threadpoolctl_1719407816649/work
tqdm @ file:///C:/b/abs_77wju137gk/croot/tqdm_1724853945787/work
traitlets==5.14.3
truststore @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/truststore_1701881385424/work
typing_extensions==4.12.2
tzdata @ file:///croot/python-tzdata_1690578112552/work
urllib3 @ file:///C:/b/abs_9a_f8h_bn2/croot/urllib3_1727769836930/work
uvicorn==0.34.0
watchfiles==1.0.4
Werkzeug==3.1.3
wheel==0.44.0
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work
wrapt==1.17.0
zstandard @ file:///C:/b/abs_82dff2ths3/croot/zstandard_1728569207191/work
